rename jsx to js and lint
